Make the cranky circuit breaker easier to reuse
This moves the `CrankyCircuitBreaker` into a top level class and
documents it so it can be used outside of testing `Aggregator`s. It
really is generally useful. It also makes constansts for the error
messages it throws. And for the error message thrown by `MockBigArrays`.
Just so it's easier to know which one throw what exception.
